# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Manager API",
        "description": "API Documentation for the Project Management System. <br> **Base URL**: /api",
        "contact": {
            "email": "admin@company.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/projects/{project}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List reviews for a project",
                "description": "Retrieve all reviews. Reviewer identity is masked for Managers/Associates.",
                "operationId": "getProjectReviews",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResource"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden. User not assigned to project."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit a review",
                "description": "Team members or Advisors can submit feedback.",
                "operationId": "createReview",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating",
                                    "content"
                                ],
                                "properties": {
                                    "rating": {
                                        "description": "1 to 5 stars",
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    },
                                    "content": {
                                        "type": "string",
                                        "minLength": 10,
                                        "example": "Great execution on the frontend deliverables."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error (e.g. Rating out of range)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get my review history",
                "description": "Personal audit log of all reviews written by the authenticated user.",
                "operationId": "getMyReviews",
                "responses": {
                    "200": {
                        "description": "List of personal reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/{review}": {
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update a review",
                "description": "Authors can edit their own reviews.",
                "operationId": "updateReview",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "content": {
                                        "type": "string",
                                        "minLength": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden. Can only edit own reviews."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "description": "Authors can delete their own reviews. Executives can delete any review.",
                "operationId": "deleteReview",
                "parameters": [
                    {
                        "name": "review",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReviewResource": {
                "title": "Review Resource",
                "description": "Review details including rating, content, and reviewer information (masked if unauthorized).",
                "properties": {
                    "id": {
                        "description": "Unique Review ID",
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "description": "Rating from 1-5",
                        "type": "integer",
                        "maximum": 5,
                        "minimum": 1,
                        "example": 5
                    },
                    "content": {
                        "description": "Review content",
                        "type": "string",
                        "example": "Great work on the backend modules."
                    },
                    "reviewer": {
                        "description": "Reviewer details. Name is masked for non-executives unless it's their own review.",
                        "properties": {
                            "id": {
                                "description": "User ID (null if anonymous)",
                                "type": "integer",
                                "example": 10,
                                "nullable": true
                            },
                            "name": {
                                "description": "Full Name (or 'Anonymous Team Member')",
                                "type": "string",
                                "example": "John Doe"
                            },
                            "role": {
                                "description": "Role name (or 'Hidden')",
                                "type": "string",
                                "example": "Manager"
                            }
                        },
                        "type": "object"
                    },
                    "project_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "project_name": {
                        "type": "string",
                        "example": "Q3 Marketing Campaign",
                        "nullable": true
                    },
                    "is_mine": {
                        "description": "True if the current user wrote this review",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}